+++ src/Version2/MultiplyOperation.java
@@ -17,7 +17,7 @@ public class MultiplyOperation extends NonTerminalExpression {
      * 将两个表达式相乘
      */
     @Override
-    public int interpreter(Context context) {
+    public double interpreter(Context context) {
         return this.e1.interpreter(context) * this.e2.interpreter(context);
     }
